+++ crates/fig_desktop/src/webview/mod.rs
@@ -25,7 +25,7 @@ use crate::{EventLoopProxy, EventLoopWindowTarget, file_watcher, local_ipc};
 
 pub const DASHBOARD_SIZE: LogicalSize<f64> = LogicalSize::new(820.0, 640.0);
 
-pub const AUTOCOMPLETE_WINDOW_TITLE: &str = "Fig Autocomplete";
+pub const AUTOCOMPLETE_WINDOW_TITLE: &str = "Fastab Autocomplete";
 
 fn map_theme(theme: &str) -> Option<TaoTheme> {
     match theme {